Fix typos (#21)
* Update the vanilla test

* Restore the confidence loss

* Fix missing parameter

* Fix the types

* Bump the .gitignore

* Fix the classes

* Fix the dtype mimatch

+++ detectinhos/vanilla.py
@@ -46,7 +46,7 @@ def to_targets(
 
     return DetectionTargets(
         boxes=np.array(bboxes),
-        classes=np.array(label_ids),
+        classes=np.array(label_ids, dtype=np.int64),
     )
